==> Muhammad Walidin_Idham Rahadian_Muhammad Fajar Maulana/1.1 - Preprocessing text.ipynb <==
By: 10117094	Muhammad Walidin
    10116476	Idham Rahadian
    10116490	Muhammad Fajar Maulana

import pandas as pd
import numpy as np
import string
import re
from collections import Counter

raw = pd.read_csv( "data/data/1.0_data_review_clean_without_stop_removal.csv")
slang_words = pd.read_csv("data/external/kbba_ID.txt",
                        sep="\t", header=None)
slang = pd.read_csv("data/external/slangword_ID.txt",
                        sep=":", header=None)
baku_words = pd.read_csv("data/external/katabaku_ID.txt",
                        sep="|", header=None)
baku_words.columns = [1,0]
slang_words = pd.concat([slang_words, slang, baku_words])

slang_words.drop_duplicates(inplace=True)
slang_words = dict(zip(slang_words[0],slang_words[1]))

print(raw.info())

raw['review'][2]

def remove_numeric(review):
    return re.sub("\d"," ", review)

raw['review'] = raw['review'].apply(remove_numeric)

raw.head()

def csv_string_to_list(csv_string):
    return csv_string[1:-1].split(",")

def string_without_quotes(word_list):
    new  = []
    for word in word_list:
        new.append(word.replace("'",""))
    return new

def delete_suffix_nya(review):
    return re.sub("(?:nya|ny)[$|\s]","",review)

raw['review'] = raw['review'].apply(delete_suffix_nya)
raw['review']

raw['review'] = raw['review'].apply(csv_string_to_list)
raw['review']

def mapping_slang_words(review):
    return [slang_words[word] if word in slang_words else word for word in review]
raw['review'] = raw['review'].apply(mapping_slang_words)

raw['review']

def remove_single_alphabet_only(review):
    return [word for word in review if word not in string.ascii_lowercase]

def remove_too_short_words(review):
    return [word for word in review if len(word) > 2]

raw['review'] = raw['review'].apply(remove_single_alphabet_only)
raw['review'] = raw['review'].apply(remove_too_short_words)

raw = raw.dropna(subset=['review'],how='all')
raw = raw[raw['review'].map(len) > 0]
raw['word_count'] = raw['review'].map(len)
def convert_list_to_string(word_list):
    return ",".join(word_list)
raw['review'] = raw['review'].apply(convert_list_to_string)
raw = raw.dropna(subset=['review'],how='all')
raw = raw[raw['review'].map(len) > 0]

raw.info()

def avg_word(words):
  return (sum(len(word) for word in words)/len(words))

raw['avg_word'] = raw['review'].apply(lambda x: avg_word(x.split(",")))

raw.to_csv("data/data/1.1_data_review_clean.csv", index=False)

most_common = pd.Series(''.join(raw['review']).split(",")).value_counts()[:10]
most_common

def add_coma(review):
    return review + ','
raw['review'] = raw['review'].apply(add_coma)

least_common = Counter("".join(raw["review"]).split(",")).most_common()[-10:]
least_common
