==> Muhammad Walidin_Idham Rahadian_Muhammad Fajar Maulana/2.0 - Splitting data dan Nilai Sentimen.ipynb <==
By: 10117094	Muhammad Walidin
    10116476	Idham Rahadian
    10116490	Muhammad Fajar Maulana

import pandas as pd
import numpy as np

raw = pd.read_csv( "data/data/1.2_data_review_cleaner_pure.csv")
raw.head()

raw['review'] = raw['review'].apply(lambda x: x.replace(","," "))
sentences = raw.loc[:,['review','rating']]
def rating_to_sentiment(rating):
    if rating > 3:
        return 1
    else:
        return 0
    
sentences['sentiment'] = sentences['rating'].apply(rating_to_sentiment)
sentences.drop(['rating'],axis=1,inplace=True)
neg = sentences[sentences['sentiment'] == 0]
pos = sentences[sentences['sentiment'] == 1]
print("All Processed Data")
print("Negative: ",neg.shape)
print("Positive: ",pos.shape)

from sklearn.model_selection import train_test_split
train, validation = train_test_split(sentences, test_size=0.2, random_state=42)
test = validation.loc[:,['review']]

train.head()

print(train.shape)

train.to_csv("data/data/train_1.csv", index=False, quoting=3)

test.head()

print(test.shape)

test.to_csv("data/data/test_1.csv", index=False, quoting=3)

validation.to_csv("data/data/validation_1.csv", index=False, quoting=3)

raw = pd.read_csv( "data/data/1.3_data_review_stemmed.csv")
raw['review'] = raw['review'].apply(lambda x: x.replace(","," "))
sentences = raw.loc[:,['review','rating']]
print(sentences)    
sentences['sentiment'] = sentences['rating'].apply(rating_to_sentiment)
sentences.drop(['rating'],axis=1,inplace=True)
neg = sentences[sentences['sentiment'] == 0]
pos = sentences[sentences['sentiment'] == 1]
print("All Processed Data")
print("Negative: ",neg.shape)
print("Positive: ",pos.shape)
sentences.to_csv("data/data/data_latih.csv", index=False, quoting=3)

train, validation = train_test_split(sentences, test_size=0.3, random_state=42)
test = validation.loc[:,['review']]

train.head()

print(train.shape)

train.to_csv("data/data/train_2.csv", index=False, quoting=3)

test.head()

print(test.shape)

test.to_csv("data/data/test_2.csv", index=False, quoting=3)

validation.to_csv("data/data/validation_2.csv", index=False, quoting=3)
