==> Muhammad Walidin_Idham Rahadian_Muhammad Fajar Maulana/4.0 - Ekstraksi fitur dan Klasifikasi SVM.ipynb <==
By: 10117094	Muhammad Walidin
    10116476	Idham Rahadian
    10116490	Muhammad Fajar Maulana

import pandas as pd
import numpy as np
import csv

# Me-Load dataset yang sudah melalui tahap preprocessing
train = pd.read_csv( "data/data/data_latih.csv")
test = pd.read_csv("data/data/test_2.csv")
validation = pd.read_csv("data/data/validation_2.csv")

train.head()

# Memisahkan data menjadi data latih dan data uji
# untuk keperluan pembangunan model
from sklearn.model_selection import train_test_split
X = train['review']
y = train['sentiment']
X_train, X_test, y_train, y_test = train_test_split(X,y,test_size=0.2, random_state=42)

# X.shape
# X_test.to_csv("1test.csv", index=False, quoting=3)

# print(X_train)
# X_train.to_csv("1fitur.csv", index=False, quoting=3)b

# Proses pembobotan atau vektorisasi
print("Creating tfidf..")
from sklearn.feature_extraction.text import TfidfVectorizer

# Inisialisasi "CountVectorizer"
vectorizer = TfidfVectorizer(analyzer = "word",
                             tokenizer = None, 
                             preprocessor = None,
                             stop_words = None, 
                             max_features = 5000) 

# Melakukan pengubahan data latih kedalam vektor fitur
%time train_data_features = vectorizer.fit_transform(X_train)

# df = pd.DataFrame(train_data_features.toarray(), columns=vectorizer.get_feature_names())
# print(df)
# df.to_csv("1vektor.csv", index=False, quoting=3)


# Mengkonversi dari list ke array, untuk memudahkan 
# ke tahap klasifikasi
train_data_features = train_data_features.toarray()

# Sama seperti pada data latih, data tes diubah kedalam
# vektor fitur selanjutnya di konversi menjadi bentuk array
test_data_features = vectorizer.transform(X_test)
test_data_features = test_data_features.toarray()

# df = pd.DataFrame(test_data_features.toarray(), columns=vectorizer.get_feature_names())
# print(df)
# df.to_csv("fitur_test.csv")

# Proses klasifikasi menggunakan SVM
from sklearn import model_selection, svm

# Inisialisasi SVM
SVM = svm.SVC(C=1.0, kernel='linear', degree=3, gamma='auto')

# Memasang data latih ke classifier
SVM.fit(train_data_features,y_train)

# Melakukan prediksi dengan membandingkan dengan data tes
pred = SVM.predict(test_data_features)


# rows = zip(X_test, pred)
# rows
# with open("1test.csv", "a") as fp:
#     wr = csv.writer(fp, dialect='excel')
#     wr.writerow(pred)

# print(X_test)

# from sklearn.metrics import classification_report
# print(classification_report(y_test, pred))

# Proses pengujian
# Import library yang dibutuhkan
from sklearn.metrics import accuracy_score,confusion_matrix
from sklearn.metrics import precision_score, recall_score, f1_score

# Melakukan proses pengujian
acc = accuracy_score(y_test, pred)
cm = confusion_matrix(y_test, pred)
print("Accuracy Score: " + str(acc))
print("Confusion Matrix: "+ str(cm))
print("Precision Score: "+ str(precision_score(y_test,pred)))
print("Recall Score: "+ str(recall_score(y_test,pred)))
print("F1 Score: "+ str(f1_score(y_test,pred)))

from sklearn import model_selection
from sklearn.model_selection import KFold, cross_val_score
from sklearn.metrics import accuracy_score, roc_auc_score, roc_curve

print(X.shape,y.shape)

X

y

import numpy as np
from numpy import mean
from numpy import std
from sklearn.model_selection import cross_val_score
# clf = svm.SVC(kernel='linear', C=1)

scores = cross_val_score(SVM, X, y, cv=5)
print('Accuracy: %.3f (%.3f)' % (mean(scores), std(scores)))

# Menguji model prediksi yang telah di buat
# dengan data masukan yang baru
tes = ['pengemasan rapi barang cepat sampai sesuai pesanan terima kasih',
       'kecewa',
       'bagus berfungsi dengan baik',
       'bungkus tidak profesional seharus memakai gelembung gelembung']
# tes_features = vectorizer.transform(tes).toarray()

est = vectorizer.transform(tes)
df = pd.DataFrame(est.toarray(), columns=vectorizer.get_feature_names())
print(df)
df.to_csv("fitur_tes.csv")
# print(tes_features)
print(SVM.predict(df))

# df = pd.DataFrame(X.toarray(), columns=tfidf2.get_feature_names())
# print(df)

# # ds = pd.read_csv("")
# print(tes)

# from csv import DictReader

# with open("data/data/test_2.csv") as f:
#     a1 = [row["review"] for row in DictReader(f)]
    
# a1

# coba = vectorizer.transform(a1).toarray()
# print(SVM.predict(coba))

# y_test.value_counts()
